Add Confrim(onRemoveLoc) finish 1

+++ js/app.controller.js
@@ -69,6 +69,8 @@ function renderLocs(locs) {
 }
 
 function onRemoveLoc(locId) {
+    const isConfrim = confirm('Are you sure?')
+    if(!isConfrim)return
     locService.remove(locId)
         .then(() => {
             flashMsg('Location removed')
